Started the Python Login Screen
There are a lot of half finished snippets of code, but I left comments where I thought it was useful. Hopefully you can use this skeleton code to finish implementing the login screen.

If you have any questions I should be on slack for most of Saturday so just message me.